Moved machine controller to machine instead
